akta-plot.py
This is a little python script that converts raw .csv data from an Äkta, like a IEX chromatography into a little plot. One can enable automated saving as .png when unhashing line 74 and further.

You need to provide the csv file path, x and y-limits as well as a title for the plot.